Conflicts:
	.gitignore
	.settings/com.genuitec.eclipse.core.prefs
	src/main/java/com/aaa/controller/ColorController.java
	src/main/java/com/aaa/controller/LoginController.java
	src/main/java/com/aaa/controller/ManagerController.java
	src/main/java/com/aaa/service/ColorService.java
	src/main/java/com/aaa/service/LibararyService.java
	src/main/java/com/aaa/service/ManagerService.java

+++ src/main/java/com/aaa/entity/EasyuiPage.java
@@ -12,13 +12,13 @@ import java.util.List;
 public class EasyuiPage {
 
 	private List rows;// 显示的数据
-	private Integer total;// 总条数
+	private Long total;// 总条数
 
 	public EasyuiPage() {
 		super();
 	}
 
-	public EasyuiPage(List rows, Integer total) {
+	public EasyuiPage(List rows, Long total) {
 		super();
 		this.rows = rows;
 		this.total = total;
@@ -32,11 +32,11 @@ public class EasyuiPage {
 		this.rows = rows;
 	}
 
-	public Integer getTotal() {
+	public Long getTotal() {
 		return total;
 	}
 
-	public void setTotal(Integer total) {
+	public void setTotal(Long total) {
 		this.total = total;
 	}
